PRO-1483: multiple-toastcontainers-cause-app-crash
